requirements.txt
discord.py==2.5.2
python-dotenv==1.1.0
pytz==2025.2
aiohttp==3.12.0
numpy==2.2.6
pymilvus==2.5.6
PyNaCl==1.5.0
scikit-learn==1.6.1
pillow==11.2.1
sentencepiece==0.2.0
neo4j==5.26.0
langchain==0.3.25
langchain-community==0.3.24
langchain-milvus==0.1.10
langchain-neo4j==0.4.0
langchain-openai==0.3.18
langchain-experimental==0.3.5rc1
beautifulsoup4==4.13.4
